Small fix with map not deploying
