+++ src/app/page.tsx
@@ -1,14 +1,11 @@
 "use client";
 
-import { MapContainer, TileLayer, Marker, Popup } from "react-leaflet";
-import { Icon } from "leaflet";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { ScrollArea } from "@/components/ui/scroll-area";
 
 import { Car, Bike } from "lucide-react";
 import { vehicles, drivers, allocations, revenueData } from "@/utils/data";
-import { Driver } from "@/types";
 import {
   Bar,
   BarChart,
@@ -17,33 +14,13 @@ import {
   YAxis,
   Tooltip,
 } from "recharts";
-import "leaflet/dist/leaflet.css";
 import VehicleInfoModal from "@/components/VehicleInfoModal";
-
-const carIcon = new Icon({
-  iconUrl: "/car-icon.png",
-  iconSize: [24, 40],
-  iconAnchor: [16, 32],
-  popupAnchor: [0, -32],
-});
-
-const bikeIcon = new Icon({
-  iconUrl: "/bike-icon.png",
-  iconSize: [24, 24],
-  iconAnchor: [16, 32],
-  popupAnchor: [0, -32],
-});
+import dynamic from "next/dynamic";
+import { getDriverForVehicle } from "@/utils/functions";
+const Map = dynamic(() => import("../components/map"), { ssr: false });
+import "leaflet/dist/leaflet.css";
 
 export default function DashboardPage() {
-  const getDriverForVehicle = (vehiclePlate: string): Driver | undefined => {
-    const allocation = allocations.find(
-      (a) => a.vehicle === vehiclePlate && a.status === "En cours"
-    );
-    return allocation
-      ? drivers.find((d) => d.id === allocation.driver)
-      : undefined;
-  };
-
   return (
     <div className="h-full space-y-4 w-full">
       <h1 className="text-3xl font-bold">Tableau de Bord IBI Fleet - Bamako</h1>
@@ -54,40 +31,7 @@ export default function DashboardPage() {
             className="p-0"
             style={{ height: "100%", width: "100%" }}
           >
-            <MapContainer
-              center={[12.6392, -8.0029]}
-              zoom={13}
-              style={{ height: "100%", width: "100%" }}
-            >
-              <TileLayer
-                url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-                attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-              />
-              {vehicles.map((vehicle) => {
-                const driver = getDriverForVehicle(vehicle.plate);
-
-                return (
-                  <Marker
-                    key={vehicle.id}
-                    position={vehicle.position}
-                    icon={vehicle.type === "car" ? carIcon : bikeIcon}
-                  >
-                    <Popup className="z-[9998]">
-                      <div className="text-center space-y-0">
-                        <h3 className="font-bold">{vehicle.plate}</h3>
-                        <p>{vehicle.type === "car" ? "Voiture" : "Moto"}</p>
-                        <p>Statut: {vehicle.status}</p>
-                        <p className="flex flex-col justify-center items-center">
-                          <span>Chauffeur: {driver?.name}</span>
-                          <span>Tel: {driver?.phone}</span>
-                        </p>
-                        <VehicleInfoModal vehicle={vehicle} driver={driver} />
-                      </div>
-                    </Popup>
-                  </Marker>
-                );
-              })}
-            </MapContainer>
+            <Map vehicles={vehicles} />
           </CardContent>
         </Card>
